pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixsorter"
version = "0.1.0"
description = "PixSorter - Lightroom-style photo grouping + best-of selection"
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "You" }]

dependencies = [
  "PySide6>=6.5",
  "numpy>=1.24",
  "pillow>=10.0",
  "scikit-learn>=1.3",
  "opencv-python>=4.8",
  "tqdm>=4.66",
]

[project.optional-dependencies]
heic = ["pillow-heif>=0.15"]
clip = ["torch", "torchvision", "open_clip_torch"]
eyes = ["mediapipe>=0.10"]

[project.scripts]
pixsorter = "pixsorter.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
pixsorter = ["assets/*"]